requirements.txt
aiogram==3.22.0
python-dotenv==1.1.1
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
openpyxl==3.1.2
